stm32_common: divide into parts for easy reuse
